OTHER APP/requirements.txt
# Audio capture (WASAPI, loopback on Windows)
soundcard>=0.4.5
# Optional: use when SoundCard crashes (set MEETINGS2_USE_PYAUDIOWPATCH=1)
pyaudiowpatch>=0.2.12
numpy>=1.24.0

# WAV I/O and resampling (capture 48k -> 16k for ASR)
soundfile>=0.12.1
scipy>=1.10.0

# NVIDIA NeMo + Parakeet ASR (local transcription)
# Install PyTorch first (with CUDA if you have GPU): https://pytorch.org
# Then: pip install nemo_toolkit['asr']
nemo_toolkit[asr]>=1.23.0
torch>=2.0.0
# Required for Python 3.12+: older lhotse (pulled in by NeMo) breaks with object.__init__()
lhotse>=1.22.0

# GUI
# (tkinter is in stdlib)
